+++ source/App.xaml
@@ -1,4 +1,4 @@
-﻿<Application x:Class="Pulse.PLMSuite.Modeller"
+﻿<Application x:Class="Pulse.PLMSuite.Modeller.App"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
             StartupUri="Views/MainWindow.xaml">
