+++ source/Modeller.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.11.35327.3
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CadModeller", "CadModeller.csproj", "{4B53C814-A2F7-4F9C-AB8E-E9238BEDC2C5}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Modeller", "Modeller.csproj", "{9467DC14-8ED9-A60A-31E2-DF274DD98CDA}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -11,10 +11,10 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{4B53C814-A2F7-4F9C-AB8E-E9238BEDC2C5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{4B53C814-A2F7-4F9C-AB8E-E9238BEDC2C5}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{4B53C814-A2F7-4F9C-AB8E-E9238BEDC2C5}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{4B53C814-A2F7-4F9C-AB8E-E9238BEDC2C5}.Release|Any CPU.Build.0 = Release|Any CPU
+		{9467DC14-8ED9-A60A-31E2-DF274DD98CDA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9467DC14-8ED9-A60A-31E2-DF274DD98CDA}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9467DC14-8ED9-A60A-31E2-DF274DD98CDA}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9467DC14-8ED9-A60A-31E2-DF274DD98CDA}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
